// nuxt-frontend/types/api.ts
export interface ApiResponse<T = any> {
  code: number
  msg: string
  data: T
  success: boolean
}